Extract formatters to a separate module
